Fix alignment of column headers and data within variant DataTables.
Issue was caused by columns not being able to auto-adjust to the correct width on creation beacuse they are hidden inside a Bootstrap collapse panel.

The solution was to bind to the shown.bs.collapse event of the variants panel and adjust the columns for each table when this event fires.