Add a field to accept_submission_requirements form where users disclose whether any samples are genetically modified.
Add a util function taking the project's reference, name of signer and a Y/N GMO flag and appends them and a timestamp to a file for that project reference in ~/.mngweb_gmo/.

Add a ChoiceField to the form with a Yes/No answer encoded as Y/N with a blank default option to flag whether any samples are genetically modified.

On form submission, call the util function to write the information to file.